+++ Libraries/Crafter.cs
@@ -6,6 +6,6 @@
         public int Control { get; init; }
         public int CP { get; init; }
         public int Level { get; init; }
-        public short[] Actions { get; init; }
+        public byte[] Actions { get; init; }
     }
 }
